backend/requirements.txt
# Django & Web Framework
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1

# Database
django-environ==0.11.2

# Security & Authentication
django-allauth==0.57.0
django-extensions==3.2.3

# API & Documentation
drf-spectacular==0.26.5
drf-spectacular-sidecar==2023.12.1

# Static Files & Frontend
whitenoise==6.6.0

# Utilities & Tools
django-filter==23.5

# Production Ready
gunicorn==21.2.0
